refactor: 분리한 상수 파일 통일 및 코드 스타일 리팩토링
